requirements.txt
ConfigArgParse==1.4
nibabel==3.2.1
pystrum==0.1
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.2
SimpleITK==2.0.2
tensorboard==2.5.0
#torch==1.9.0
#torchaudio==0.9.0
#torchvision==0.10.0
torchsummary==1.5.1
torchinfo==0.1.1
tqdm==4.60.0
h5py==2.10.0
matplotlib==3.3.4
marshmallow==3.0.0rc5
ml-collections==0.1.0
numpy==1.19.5
Pillow==8.2.0
polyaxon-cli==0.6.1
polyaxon-client==0.6.1
polyaxon-deploy==0.6.1
polyaxon-helper==0.5.0
polyaxon-schemas==0.6.1
psutil==5.8.0
monai==0.6.0
optuna==2.8.0
joblib==1.0.1
PyYAML==5.4.1